Update GlobalTurnIn.lua
Migrated a function out of the Shop loop (not sure why it was there but let's see what breaks.

Rewrote the logic for open shop to also include raid tier (so it should back out and go through each one. Still need to edit to change the loop logic for opening.